requirements.txt
Flask==2.3.3
pyserial==3.5
pymodbus==2.5.3
RPi.GPIO==0.7.1a